fix: show anchors in collection link field
