+++ app/(builder)/ycode/components/LinkSettings.tsx
@@ -42,7 +42,7 @@ import { useComponentsStore } from '@/stores/useComponentsStore';
 import { ASSET_CATEGORIES, getAssetIcon } from '@/lib/asset-utils';
 import { toast } from 'sonner';
 import { collectionsApi, pagesApi } from '@/lib/api';
-import { getLayerIcon, getLayerName, canLayerHaveLink, getCollectionVariable } from '@/lib/layer-utils';
+import { getLayerIcon, getLayerName, canLayerHaveLink, getCollectionVariable, findLayersWithAnchorId } from '@/lib/layer-utils';
 import { Empty, EmptyDescription, EmptyTitle } from '@/components/ui/empty';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
 import PageSelector from './PageSelector';
@@ -190,50 +190,23 @@ export default function LinkSettings(props: LinkSettingsProps) {
     return pages.find((p) => p.id === pageId) || null;
   }, [pageId, pages]);
 
-  // Flatten layers and find all layers with a custom ID (settings.id takes priority over attributes.id)
-  const findLayersWithId = useCallback((layers: Layer[]): Array<{ layer: Layer; id: string }> => {
-    const result: Array<{ layer: Layer; id: string }> = [];
-    const stack: Layer[] = [...layers];
-
-    while (stack.length > 0) {
-      const layer = stack.pop()!;
-
-      const layerId = layer.settings?.id || layer.attributes?.id;
-      if (layerId) {
-        result.push({ layer, id: layerId });
-      }
-
-      if (layer.children) {
-        stack.push(...layer.children);
-      }
-    }
-
-    return result;
-  }, []);
-
   // Get layers for anchor selection based on link type
   const anchorLayers = useMemo(() => {
     let targetPageId: string | null = null;
 
     if (linkType === 'page' && pageId) {
-      // For page links, use the selected page
       targetPageId = pageId;
     } else if (linkType === 'url' && currentPageId) {
-      // For URL links, use the current page
       targetPageId = currentPageId;
     }
 
-    if (!targetPageId) {
-      return [];
-    }
+    if (!targetPageId) return [];
 
     const draft = draftsByPageId[targetPageId];
-    if (!draft || !draft.layers) {
-      return [];
-    }
+    if (!draft || !draft.layers) return [];
 
-    return findLayersWithId(draft.layers);
-  }, [linkType, pageId, currentPageId, draftsByPageId, findLayersWithId]);
+    return findLayersWithAnchorId(draft.layers);
+  }, [linkType, pageId, currentPageId, draftsByPageId]);
 
   // Check if selected page is dynamic
   const isDynamicPage = selectedPage?.is_dynamic || false;
